% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadMatch.R
\name{downloadMatch}
\alias{downloadMatch}
\title{Download data from a single match}
\usage{
downloadMatch(comp_id, round_id, game_id)
}
\arguments{
\item{comp_id}{A string identifying which season the game is
in. \code{comp_id} is different depending on regular season or finals.}

\item{round_id}{An integer identifying which round the game is in. Finals
reset round number to 1.}

\item{game_id}{An integer indentifying which game in the round to
download. There are four games per round in the regular season, two games
in the semi finals, one game for the prelim, and one grand final.}
}
\value{
A list containing game and player data for the match.
}
\description{
\code{downloadMatch} downloads match and player data for a single match.
}
\examples{
\dontrun{
downloadMatch("10083", 1, 1)
}

}
