% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ladders.R
\name{ladders}
\alias{ladders}
\alias{matchResults}
\title{Calculates ladder positions}
\usage{
ladders(df, round_num = NULL, game_num = NULL, old_system = FALSE)

matchResults(df)
}
\arguments{
\item{df}{Data frame containing season match statistics.}

\item{round_num}{Round at which to calculate ladder positions. Optional.}

\item{game_num}{Game at which to calculate ladder positions. Optional.}

\item{old_system}{Logical. Whether to sort by the old scoring system
(defaults to FALSE).}
}
\value{
Data frame containing the ladder position of all teams. If round and
    game are not supplied, the ladder position is calculated using all match
    data present in the \code{df} supplied.
}
\description{
\code{ladders} calculates ladder positions at the end of a match.
}
