% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchPoints.R
\name{matchPoints}
\alias{matchPoints}
\title{Calculates the total goals of the match}
\usage{
matchPoints(df)
}
\arguments{
\item{data}{Match data.}
}
\value{
A data frame containing the final scores, and points for the ladder.
}
\description{
\code{matchPoints} calculates final match goals and score difference.
}
