% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchPoints.R
\name{matchPoints_pre_2020}
\alias{matchPoints_pre_2020}
\title{Calculates the total goals of the match (pre 2020 season)}
\usage{
matchPoints_pre_2020(df)
}
\arguments{
\item{df}{Match data.}
}
\value{
A data frame containing the final scores, and points for the ladder.
}
\description{
\code{matchPoints_pre_2020} calculates final match goals and score
difference, for seasons pre-2020.
}
