% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{players_2017}
\alias{players_2017}
\title{Season 2017 player data.}
\format{
A data frame with 163336 rows and 8 variables:
\describe{
  \item{playerId}{Unique player number}
  \item{period}{Which period the statistic is measured in}
  \item{squadId}{Unique squad number}
  \item{shortDisplayName}{surname, firstname}
  \item{firstname}{Player firstname}
  \item{surname}{Player surname}
  \item{squadName}{Full squad name}
  \item{stat}{Statistic measured during the match}
  \item{value}{Value of the statistic}
  \item{round}{Round number of the match}
  \item{game}{Game number of the match}
}
}
\usage{
players_2017
}
\description{
A dataset containing player statistics for all home and away and finals
series matches from the 2017 super netball season, by period.
}
\keyword{datasets}
