% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{round5_game3}
\alias{round5_game3}
\title{Match and player statistics from round 5, game 3, season 2017.}
\format{
A list.
}
\usage{
round5_game3
}
\description{
A list containing detailed match and player statistics, as obtained using the
\code{downloadMatch} function.
}
\keyword{datasets}
