% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{season_2017}
\alias{season_2017}
\title{Season 2017 match data.}
\format{A data frame with 15360 rows and 8 variables:
\describe{
  \item{squadId}{Unique squad number}
  \item{squadName}{Full squad name}
  \item{squadNickname}{Squad nickname}
  \item{squadCode}{Short code for quad}
  \item{stat}{Statistic measured during the match}
  \item{value}{Value of the statistic}
  \item{round}{Round number of the match}
  \item{game}{Game number of the match}
}}
\usage{
season_2017
}
\description{
A dataset containing match statistics for all home and away and finals series
matches from the 2017 super netball season, by period.
}
\keyword{datasets}
