% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinySuperNetballR.R
\name{shinySuperNetballR}
\alias{shinySuperNetballR}
\title{Runs the demo shiny app}
\usage{
shinySuperNetballR()
}
\value{
Runs a shiny app
}
\description{
\code{shinySuperNetballR} Runs the demo shiny app to compare Super Netball
statistics between teams.
}
