% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{team_colours}
\alias{team_colours}
\title{Team colours.}
\format{
A data frame with 8 rows and 3 variables:
\describe{
  \item{squadName}{Full squad name}
  \item{squadId}{Unique squad number}
  \item{squadColour}{Hex-coded team colour}
}
}
\usage{
team_colours
}
\description{
A dataset containing hex-coded team colours for each team.
}
\keyword{datasets}
