% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidiers.R
\name{tidyMatch}
\alias{tidyMatch}
\title{Takes a downloaded match list and tidies the match statistics.}
\usage{
tidyMatch(match)
}
\arguments{
\item{match}{List of match details.}
}
\value{
A tidy dataframe containing match statistics.
}
\description{
\code{tidyMatch} Takes the downloaded match list, and tidies match statistics
in preparation for further analysis.
}
