% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidiers.R
\name{tidyPlayers}
\alias{tidyPlayers}
\title{Takes a downloaded match list and tidies the player statistics.}
\usage{
tidyPlayers(match)
}
\arguments{
\item{match}{List of match details.}
}
\value{
A tidy dataframe containing player statistics.
}
\description{
\code{tidyPlayers} Takes the downloaded match list, and tidies player
statistics in preparation for further analysis.
}
